C11-159: relocate v1+v2 dispatch layer into SocketDispatch.swift
Removes every socket dispatch switch from TerminalController.swift (DX-1):
processCommand (v1 switch cmd, 112 cases), processV2Command, the off-main
socket-worker fast paths (socketWorkerV1Response switch head, socketWorkerV2Response
switch request.method), asyncAck/policy routing, the pure parser helpers, the
telemetry worker variants, and the v2DispatchExtracted per-domain router all move
verbatim into Sources/SocketHandlers/SocketDispatch.swift. Threading tiers
preserved: nonisolated members stay nonisolated (off-main); processCommand/
processV2Command stay main-actor. The vestigial empty v2 switch is collapsed to a
direct v2DispatchExtracted call. v1 command handlers stay in TerminalController,
widened to internal so the relocated switch can call them.
TerminalController.swift: 9676 -> 8774 LOC. No dispatch switch remains in it.
